rename 'input-types' to 'inputs-types' + fix imports
